Move HexMap.py to Outputs folder
